// src/app/modules/transportation/models/vehicle-information-view.ts
import { AttachmentViewModel } from 'app/modules/academics/student-admissions/model/attachment-view-model';

/**
 * Transport
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

/**
 * Gets or Sets VehicleInformationView
 */
export interface VehicleInformationView {
    /**
        * Gets Or Sets
        */
    id?: string;
    /**
     * Gets Or Sets
     */
    vehicleName?: string;
    /**
     * Gets Or Sets
     */
    model?: string;
    /**
     * Gets Or Sets
     */
    manufactureYear?: Date;
    /**
     * Gets Or Sets
     */
    regNumber?: string;
    /**
     * Gets Or Sets
     */
    engineNumber?: string;
    /**
     * Gets Or Sets
     */
    chassisNumber?: string;
    /**
     * Gets Or Sets
     */
    maxCapacity?: number;

    /**
    * Gets Or Sets
    */
    manufacturer?: string;
    /**
     * Gets Or Sets
     */
    registrationCertAttachment?: AttachmentViewModel;
    /**
   * Gets Or Sets
   */
    photoAttachment?: AttachmentViewModel;
}